Allow to lock segments
SegmentsModule

- Added `locked` field to `segments` table.
- Updated segment's form to disallow update of locked segment.
- Added `setLock()` method to lock / unlock segment.

SegmentRepository methods `add()`, `update()` and `upsert()` cannot change
segment's locked property and locked segment cannot be updated. It's
resposibility of caller to decide if update of locked segment should be
allowed (and caller needs to unlock it). Special method `setLock()` for
locking / unlocking of segment has to be called.

OnboardingModule

- Added `locked` field set to true to all segments seeded for
  onboarding goals. These segments are generated, we don't want to
  allow use to change it.

remp/crm#1763