Add UserOnboardingGoalEventsHandler to update user in cached segment
Handler listens to user onboarding goal events and adds or removes user
entry in REMP/Campaign's cached CRM segment. This is done through API
call. Change is only temporary; next scheduled cache job within Campaign
will update list with current values (based on call to CRM segments).

We need to reflect changes to user's onboarding goals in real time so we
don't show campaigns after user's onboarding goal was completed / timed
out.

Warning: `UserOnboardingGoalEventsHandler` uses API client from
RempCampaignModule which is not mandatory module. So handler cannot be
part of OnboardingModule and it lives in RempCampaignModule. If you want
to use this handler to update cached onboarding segments in Campaign,
both modules have to be installed and enabled and this handlers has to
be registered as lister in your module.

Note: If Campaign's API for overriding segment cache returns response code
404 NotFound, no active campaign with provided segment was found. This is not
an error and is not logged as one. If segment is not actively used, no need
to override it's cache.

remp/crm#1765